requirements.txt
# python=3.10
vllm
transformers
optimum
accelerate==0.21.0
peft==0.5.0
scikit-learn==1.3.0
spacy==3.6.1
python-Levenshtein
nltk
tqdm
tensorboard